requirements.txt
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
plotly==5.22.0
scanpy==1.10.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.1
umap-learn==0.5.5
fastparquet==2023.10.1